accessible/tests/mochitest/tree/test_aria_grid.html

<!DOCTYPE html>
<html>
<head>
  <title>HTML table tests</title>
  <link rel="stylesheet" type="text/css"
        href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <script type="application/javascript"
          src="../common.js"></script>
  <script type="application/javascript"
          src="../role.js"></script>

  <script type="application/javascript">
    function doTest()
    {
      //////////////////////////////////////////////////////////////////////////
      // grid having rowgroups

      var accTree =
        { TABLE: [
          { GROUPING: [
            { ROW: [
              { GRID_CELL: [
                { TEXT_LEAF: [ ] }
              ] }
            ] }
          ] },
        ] };

      testAccessibleTree("grid", accTree);


      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  </script>
</head>
<body>

  <a target="_blank"
     title="Support ARIA role rowgroup"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=525909">
    Mozilla Bug 525909
  </a>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test">
  </pre>

  <div id="grid" role="grid">
    <div role="rowgroup">
      <div role="row">
        <div role="gridcell">cell</div>
      </div>
    </div>
  </div>
</body>
</html>
